:hammer: Rename parse_time module to time_tools
Also add json_time.

The module was already kinda badly-named anyway, and adding json_time just
made it worse. So this adds json_time *and* fixes the naming.
